Serializador timestamp (uint64_t)

+++ Librerias/serializador/estructuras.c
@@ -1,6 +1,6 @@
 #include "estructuras.h"
 
-registro_tad* new_registro_tad(uint32_t timestamp, uint32_t key, char* value){
+registro_tad* new_registro_tad(uint64_t timestamp, uint32_t key, char* value){
     registro_tad* registro = malloc(sizeof(registro_tad));
     registro->timestamp = timestamp;
     registro->key = key;
